Make sign-in and sign-up server components, separate sign-up form, use axios instead of axiosInstance, and adjust SVG components
