+++ components/features/auth/SigninForm.tsx
@@ -1,0 +1,121 @@
+"use client";
+
+import axios from "@/lib/axiosInstance";
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { z } from "zod";
+import { AxiosError } from "axios";
+import { useRouter } from "next/navigation";
+import { setCookie } from "cookies-next/client";
+import { useState } from "react";
+import { signinFormSchema } from "@/lib/validations";
+import { Button } from "@/components/ui/button";
+import {
+  Form,
+  FormMessage,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormControl,
+} from "@/components/ui/form";
+import { EyeOffIcon, EyeIcon, LoaderCircle as SpinnerIcon } from "lucide-react";
+import { Input } from "@/components/ui/input";
+import { ResetPassword } from "./ResetPassword";
+
+type formDataType = z.infer<typeof signinFormSchema>;
+
+export default function SigninForm() {
+  const router = useRouter();
+  const [isPasswordVisible, setIsPasswordVisible] = useState(false);
+  const form = useForm<formDataType>({
+    resolver: zodResolver(signinFormSchema),
+    defaultValues: {
+      email: "",
+      password: "",
+    },
+  });
+
+  async function onSubmit(formData: formDataType) {
+    try {
+      const res = await axios.post("/api/sign-in", formData);
+      setCookie("token", res.data.token);
+      router.push("/console");
+    } catch (err) {
+      const error = err as AxiosError;
+      console.error("Could not sign in: ", error);
+      form.setError("root", { message: error.response?.data as string });
+    }
+  }
+  return (
+    <Form {...form}>
+      <form className="w-full space-y-4" onSubmit={form.handleSubmit(onSubmit)}>
+        {!!form.formState.errors && (
+          <FormMessage>{form.formState.errors.root?.message}</FormMessage>
+        )}
+
+        <FormField
+          name="email"
+          control={form.control}
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-muted-foreground">Email</FormLabel>
+              <FormControl>
+                <Input type="email" placeholder="Email" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          name="password"
+          control={form.control}
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-muted-foreground">Password</FormLabel>
+              <FormControl>
+                <div className="relative">
+                  <Input
+                    type={isPasswordVisible ? "text" : "password"}
+                    placeholder="Password"
+                    {...field}
+                  />
+
+                  <button
+                    className="absolute right-1 top-0 p-2 text-muted-foreground"
+                    onClick={() => setIsPasswordVisible((prev) => !prev)}
+                    type="button"
+                  >
+                    {isPasswordVisible ? (
+                      <EyeOffIcon size={25} />
+                    ) : (
+                      <EyeIcon size={25} />
+                    )}
+                  </button>
+                </div>
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <ResetPassword />
+
+        <Button
+          onClick={form.handleSubmit(onSubmit)}
+          disabled={
+            form.formState.isSubmitting ||
+            !!form.formState.errors.email ||
+            !!form.formState.errors.password
+          }
+          className="mt-5 w-full rounded-full disabled:cursor-not-allowed"
+        >
+          {form.formState.isSubmitting ? (
+            <SpinnerIcon className="animate-spin" size={25} />
+          ) : (
+            <span>Sign in</span>
+          )}
+        </Button>
+      </form>
+    </Form>
+  );
+}
